Corrected bug with the tears. now the end of their lives is determinated by a new component called LifeTimeComponen. Added also the tear rate to the player. added the comments on the xml. added function to the clas Position. And other changes

+++ src/model/component/TearAIComponent.java
@@ -1,6 +1,8 @@
 package model.component;
 
 import com.google.common.eventbus.Subscribe;
+
+import model.component.mentality.AbstractMentalityComponent;
 import model.entity.Entity;
 import model.enumeration.BasicStatusEnum;
 import model.events.CollisionEvent;
@@ -12,75 +14,26 @@ import util.EventListener;
  */
 public class TearAIComponent extends AbstractComponent<TearAIComponent> {
 
-    private static final double DEFAULT_TEAR_LIFETIME = 200;
-    private final double lifetime;
-    private final int angle;
-    private double time;
-
-    /**
-     * @param entity this entity
-     */
-    public TearAIComponent(final Entity entity) {
-        this(entity, 0, DEFAULT_TEAR_LIFETIME);
-    }
-
-    /**
-     * @param entity this entity
-     * @param angle direction angle of the tear
-     */
-    public TearAIComponent(final Entity entity, final int angle) {
-        this(entity, angle, DEFAULT_TEAR_LIFETIME);
-    }
-
-    /**
-     * @param entity this entity
-     * @param lifetime time before the tear disappears
-     */
-    public TearAIComponent(final Entity entity, final double lifetime) {
-        this(entity, 0, lifetime);
-    }
-
     /**
      * the listener of this component handles the disappearance of the entity when
      * it collides with something.
      * @param entity this entity
-     * @param angle direction angle of the tear
-     * @param lifetime time before the tear disappears
      */
-    public TearAIComponent(final Entity entity, final int angle, final double lifetime) {
+    public TearAIComponent(final Entity entity) {
         super(entity);
-        this.angle = angle;
-        this.lifetime = lifetime;
 
         this.registerListener(new EventListener<CollisionEvent>() {
             @Subscribe
             @Override
             public void listenEvent(final CollisionEvent event) {
-                getEntity().getRoom().deleteEntity(getEntity());
+                if (getEntity().hasComponent(AbstractMentalityComponent.class)
+                        && event.getSourceEntity().hasComponent(AbstractMentalityComponent.class)
+                        && getEntity().getComponent(AbstractMentalityComponent.class)
+                        .get().canCollide(event.getSourceEntity().getComponent(AbstractMentalityComponent.class).get().getClass())) {
+                    getEntity().getStatusComponent().setStatus(BasicStatusEnum.DISAPPEAR);
+                    getEntity().getRoom().deleteEntity(getEntity());
+                }
             }
         });
-    }
-
-    /**
-     * 
-     * @return direction
-     */
-    public int getAngle() {
-        return this.angle;
-    }
-
-    /**
-     * {@inheritDoc}
-     * Checks if the lifetime of the tear is expired. If so, makes it disappear.
-     */
-    @Override
-    public void update(final Double deltaTime) {
-        time += deltaTime;
-        if (time > lifetime) {
-            getEntity().getRoom().deleteEntity(this.getEntity());
-            getEntity().getStatusComponent().setStatus(BasicStatusEnum.DISAPPEAR);
-        } else {
-            super.getEntity().getComponent(MoveComponent.class).get().move(angle);
-        }
     }
 }
